+++ dart/test/sentry_test.dart
@@ -185,6 +185,7 @@ void main() {
     await Sentry.init((options) {
       options.dsn = fakeDsn;
       expect(options.environment, 'debug');
+      expect(options.debug, true);
     }, options: sentryOptions);
   });
 
@@ -194,6 +195,7 @@ void main() {
     await Sentry.init((options) {
       options.dsn = fakeDsn;
       expect(options.environment, 'profile');
+      expect(options.debug, false);
     }, options: sentryOptions);
   });
 
@@ -202,7 +204,8 @@ void main() {
       ..platformChecker = FakePlatformChecker.releaseMode();
     await Sentry.init((options) {
       options.dsn = fakeDsn;
-      expect(options.environment, defaultEnvironment);
+      expect(options.environment, 'production');
+      expect(options.debug, false);
     }, options: sentryOptions);
   });
 }