+++ CHANGELOG.md
@@ -7,6 +7,8 @@
 * Bump: sentry-cocoa to v6.2.1 (#360)
 * Feat: Migration from `package_info` to `package_info_plus` plugin (#370)
 * Fix: missing event.origin and event.environment tags for non-native events on iOS (#369)
+* Fix: Set `SentryOptions.debug` in `sentry` (#376)
+* Fix: Read all environment variables in `sentry` (#375)
 
 ## Breaking Changes: